+++ crates/matrix-sdk-ui/src/timeline/futures.rs
@@ -1,4 +1,4 @@
-use std::{fs, future::IntoFuture, path::PathBuf};
+use std::future::IntoFuture;
 
 use eyeball::{SharedObservable, Subscriber};
 use matrix_sdk::{attachment::AttachmentConfig, TransmissionProgress};
@@ -6,11 +6,11 @@ use matrix_sdk_base::boxed_into_future;
 use mime::Mime;
 use tracing::{Instrument as _, Span};
 
-use super::{Error, Timeline};
+use super::{AttachmentSource, Error, Timeline};
 
 pub struct SendAttachment<'a> {
     timeline: &'a Timeline,
-    path: PathBuf,
+    source: AttachmentSource,
     mime_type: Mime,
     config: AttachmentConfig,
     tracing_span: Span,
@@ -21,13 +21,13 @@ pub struct SendAttachment<'a> {
 impl<'a> SendAttachment<'a> {
     pub(crate) fn new(
         timeline: &'a Timeline,
-        path: PathBuf,
+        source: AttachmentSource,
         mime_type: Mime,
         config: AttachmentConfig,
     ) -> Self {
         Self {
             timeline,
-            path,
+            source,
             mime_type,
             config,
             tracing_span: Span::current(),
@@ -61,16 +61,18 @@ impl<'a> IntoFuture for SendAttachment<'a> {
     boxed_into_future!(extra_bounds: 'a);
 
     fn into_future(self) -> Self::IntoFuture {
-        let Self { timeline, path, mime_type, config, tracing_span, use_send_queue, send_progress } =
-            self;
+        let Self {
+            timeline,
+            source,
+            mime_type,
+            config,
+            tracing_span,
+            use_send_queue,
+            send_progress,
+        } = self;
 
         let fut = async move {
-            let filename = path
-                .file_name()
-                .ok_or(Error::InvalidAttachmentFileName)?
-                .to_str()
-                .ok_or(Error::InvalidAttachmentFileName)?;
-            let data = fs::read(&path).map_err(|_| Error::InvalidAttachmentData)?;
+            let (data, filename) = source.try_into_bytes_and_filename()?;
 
             if use_send_queue {
                 let send_queue = timeline.room().send_queue();