+++ crates/matrix-sdk/CHANGELOG.md
@@ -15,6 +15,8 @@ All notable changes to this project will be documented in this file.
 - [**breaking**] Remove the `AttachmentConfig::with_thumbnail()` constructor and
   replace it with the `AttachmentConfig::thumbnail()` builder method. You should
   call `AttachmentConfig::new().thumbnail(thumbnail)` now instead.
+- [**breaking**] `Room::send_attachment()` and `RoomSendQueue::send_attachment()`
+  now take any type that implements `Into<String>` for the filename.
 
 ## [0.9.0] - 2024-12-18